Split TransferForm and TransferTable in demo-transfer-dashboard
