+++ demo-transfer-dashboard/src/App.jsx
@@ -7,6 +7,7 @@ import {
 } from '@cdssnc/gcds-components-react';
 
 import TransferForm from './Components/TransferForm.jsx';
+import TransferTable from './Components/TransferTable.jsx';
 import WorkInProgressAlert from './Components/WorkInProgressAlert.jsx';
 
 import { get_default_pt } from './pt_utils.js';
@@ -27,6 +28,7 @@ function App() {
           PT-to-PT MMR Immunization Record Transfer
         </GcdsHeading>
         <TransferForm defaultPT={default_pt} />
+        <TransferTable defaultPT={default_pt} />
         <GcdsDateModified> {process.env.BUILD_DATE}</GcdsDateModified>
       </main>
